'default' target
This adds code to interpret the string `'default'` as the target that
corresponds to the current host (in this case, the “Default Host”
according to `rustup show`).

This switches to not purging by default, since that is what removes the
target installed when installing the toolchain.

This adds purge support to `rustup::global`.