Clarify how `manage_*` function change state
Previously, this reloaded the toolchains in the `manage_targets`
function, but it was not obvious that it would have that side effect.

This moves the state reset out of the function and adds comments
mentioning that the `manage_*` functions do not update internal state.
This will be helpful for managing components.