Fix subresource identity function
The identity function is supposed to uniquely identify a subresource in
a set for the purpose of detecting defaults. At its most basic, it’s
supposed to just just remove the `ensure` attribute and compute the
hash so you can’t have two otherwise identical subresources with
different values for `ensure`.

Previously it stripped an attribute called `absent` rather than the
`ensure` attribute.