Clean up better after individual acceptance tests
Previously, running the acceptance tests twice on the same VM or
container would cause a couple of test failures. This fixes it, though
I’m not entirely sure how, since one of the errors was that the user
existed at the beginning of the test… and the tests already deleted the
user.