Fix display of changes to subresources
When a parameter changes, Puppet displays a change message with the old
and new values. Switching to the subresource collection classes on the
provider broke this display — the old value was displayed as an opaque
ruby object.

This restores the change display to show the interior attributes of all
subresourcces.