Fix old values for subresources in change display
When a property changes on a resource, Puppet displays a change message
with the old and new values. Previously, the value displayed as “old”
was actually just the new values.

This fixes the problem.

It appears that Puppet may expect the collection object to be replaced
rather than change.

Fixes #61 — Change messages for subresources don’t show old values